services/movies/requirements.txt
flask==3.0.0
flask-restx==1.3.0
psycopg2-binary==2.9.9
neo4j==5.20.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
